change timings of setting slot and indicator and prepare cmd for nArbiSlots

+++ src/wifi/model/ctrl-headers.h
@@ -1262,7 +1262,7 @@ private:
   bool m_mbtaIndicator;   //!< Indicator of whether we use MBTA(1) or not(0)
   uint8_t m_arbitrationSlot; //!< Number of Arbitration Slots. Added by Ryu 2022/10/5
   bool m_damsIndicator;   //!< Indicator of whether we use DAMS(1) or not(0)
-  uint8_t m_damsBoundary; //!< Boundary of frame length between UORA and UONRA
+  uint16_t m_damsBoundary; //!< Boundary of frame length between UORA and UONRA
   /**
    * List of User Info fields
    */
